Moved common types to cfg types
