+++ libs/types/src/lib.rs
@@ -17,11 +17,13 @@
 ///! Common-types of the Centrifuge chain.
 pub mod adjustments;
 pub mod consts;
+pub mod domain_address;
 pub mod epoch;
 pub mod fee_keys;
 pub mod fixed_point;
 pub mod ids;
 pub mod investments;
+pub mod locations;
 pub mod orders;
 pub mod permissions;
 pub mod time;
